Got the remove button working when no choice is bade on remove cheese and remove from menu. Learned new requestparam (required=false) as part of it. Then using an if statement.

+++ src/main/java/org/launchcode/controllers/CheeseController.java
@@ -69,12 +69,14 @@ public class CheeseController {
     }
 
     @RequestMapping(value = "remove", method = RequestMethod.POST)
-    public String processRemoveCheeseForm(@RequestParam int[] cheeseIds) {
-
+    public String processRemoveCheeseForm(@RequestParam (required=false) int[] cheeseIds) {
+        //added the (required=false) as part of the if update mentioned below
+//updated on my own projuect added iff statement so hitting remove with no selection wouldnt break it
+        if (cheeseIds != null) {
         for (int cheeseId : cheeseIds) {
             System.out.println("remove cheese  " + cheeseIds);
             cheeseDao.delete(cheeseId);
-        }
+        }}
 
         return "redirect:";
     }
